+++ src/player.hpp
@@ -72,8 +72,8 @@ private:
 	uint64_t prev_move_time_ms_;
 
 	m_Vec3 build_pos_;
-	short discret_build_pos_[3]; // World space build position.
-	short destroy_pos_[3];
+	int discret_build_pos_[3]; // World space build position.
+	int destroy_pos_[3];
 	h_Direction build_direction_; // Firection of build side. Unknown, if build block does not exist
 	h_BlockType build_block_;
