+++ src/world_lighting.cpp
@@ -16,24 +16,24 @@ static const unsigned int g_vertex_light_div_table[]=
 	g_vertex_light_div_table_multiplier * (13+6)/6,
 };
 
-unsigned char h_World::SunLightLevel( short x, short y, short z ) const
+unsigned char h_World::SunLightLevel( int x, int y, int z ) const
 {
 	return
 		GetChunk( x >> H_CHUNK_WIDTH_LOG2, y>> H_CHUNK_WIDTH_LOG2 )->
 		SunLightLevel( x& (H_CHUNK_WIDTH-1), y& (H_CHUNK_WIDTH-1), z );
 }
-unsigned char h_World::FireLightLevel( short x, short y, short z ) const
+unsigned char h_World::FireLightLevel( int x, int y, int z ) const
 {
 	return
 	GetChunk( x >> H_CHUNK_WIDTH_LOG2, y>> H_CHUNK_WIDTH_LOG2 )->
 	FireLightLevel( x& (H_CHUNK_WIDTH-1), y& (H_CHUNK_WIDTH-1), z );
 }
 
-void h_World::GetForwardVertexLight( short x, short y, short z, unsigned char* out_light ) const
+void h_World::GetForwardVertexLight( int x, int y, int z, unsigned char* out_light ) const
 {
-	unsigned short block_count= 0;
+	unsigned int block_count= 0;
 	unsigned char light[2]= { 0, 0 };
-	short  x1, y1;
+	int  x1, y1;
 	const h_Chunk* ch;
 	unsigned int addr;
 
@@ -95,11 +95,11 @@ void h_World::GetForwardVertexLight( short x, short y, short z, unsigned char* o
 	out_light[1]= ( ( (unsigned int)light[1] ) * g_vertex_light_div_table[ block_count ] )>> g_vertex_light_div_shift;
 }
 
-void h_World::GetBackVertexLight( short x, short y, short z, unsigned char* out_light ) const
+void h_World::GetBackVertexLight( int x, int y, int z, unsigned char* out_light ) const
 {
-	unsigned short block_count= 0;
+	unsigned int block_count= 0;
 	unsigned char light[2]= { 0, 0 };
-	short  x1, y1;
+	int  x1, y1;
 	const h_Chunk* ch;
 	unsigned int addr;
 
@@ -161,13 +161,13 @@ void h_World::GetBackVertexLight( short x, short y, short z, unsigned char* out_
 	out_light[1]= ( ( (unsigned int)light[1] ) * g_vertex_light_div_table[ block_count ] )>> g_vertex_light_div_shift;
 }
 
-void h_World::SetSunLightLevel( short x, short y, short z, unsigned char l )
+void h_World::SetSunLightLevel( int x, int y, int z, unsigned char l )
 {
 	GetChunk( x >> H_CHUNK_WIDTH_LOG2, y>> H_CHUNK_WIDTH_LOG2 )->
 	sun_light_map_[ BlockAddr(  x& (H_CHUNK_WIDTH-1), y& (H_CHUNK_WIDTH-1), z ) ]= l;
 }
 
-void h_World::SetFireLightLevel( short x, short y, short z, unsigned char l )
+void h_World::SetFireLightLevel( int x, int y, int z, unsigned char l )
 {
 	GetChunk( x >> H_CHUNK_WIDTH_LOG2, y>> H_CHUNK_WIDTH_LOG2 )->
 	fire_light_map_[ BlockAddr(  x& (H_CHUNK_WIDTH-1), y& (H_CHUNK_WIDTH-1), z ) ]= l;
@@ -180,10 +180,10 @@ void h_World::LightWorld()
 	for( unsigned int j= 1; j< chunk_number_y_ - 1; j++ )
 	{
 		h_Chunk* ch= GetChunk( i, j );
-		short X= i * H_CHUNK_WIDTH, Y= j * H_CHUNK_WIDTH;
-		for( short x= 0; x< H_CHUNK_WIDTH; x++ )
-		for( short y= 0; y< H_CHUNK_WIDTH; y++ )
-		for( short z= 1; z< H_CHUNK_HEIGHT - 1; z++ )
+		int X= i * H_CHUNK_WIDTH, Y= j * H_CHUNK_WIDTH;
+		for( int x= 0; x< H_CHUNK_WIDTH; x++ )
+		for( int y= 0; y< H_CHUNK_WIDTH; y++ )
+		for( int z= 1; z< H_CHUNK_HEIGHT - 1; z++ )
 			AddSunLight_r( X+x, Y+y, z, ch->SunLightLevel(x,y,z) );
 
 		const std::vector< h_LightSource* >& light_sources= ch->GetLightSourceList();
@@ -196,10 +196,10 @@ void h_World::LightWorld()
 	for( unsigned int i= 0; i< chunk_number_x_; i++ )
 	{
 		h_Chunk* ch= GetChunk( i, k * (chunk_number_y_-1) );
-		short X= i * H_CHUNK_WIDTH, Y=  H_CHUNK_WIDTH * k * (chunk_number_y_-1) ;
-		for( short x= 0; x< H_CHUNK_WIDTH; x++ )
-		for( short y= 0; y< H_CHUNK_WIDTH; y++ )
-		for( short z= 1; z< H_CHUNK_HEIGHT-1; z++ )
+		int X= i * H_CHUNK_WIDTH, Y=  H_CHUNK_WIDTH * k * (chunk_number_y_-1) ;
+		for( int x= 0; x< H_CHUNK_WIDTH; x++ )
+		for( int y= 0; y< H_CHUNK_WIDTH; y++ )
+		for( int z= 1; z< H_CHUNK_HEIGHT-1; z++ )
 			AddSunLightSafe_r( X+x, Y+y, z, ch->SunLightLevel(x,y,z) );
 
 		const std::vector< h_LightSource* >& light_sources= ch->GetLightSourceList();
@@ -212,10 +212,10 @@ void h_World::LightWorld()
 	for( unsigned int j= 0; j< chunk_number_y_; j++ )
 	{
 		h_Chunk* ch= GetChunk( k*( chunk_number_x_ - 1), j );
-		short X= k*( chunk_number_x_ - 1) * H_CHUNK_WIDTH, Y=  j * H_CHUNK_WIDTH;
-		for( short x= 0; x< H_CHUNK_WIDTH; x++ )
-		for( short y= 0; y< H_CHUNK_WIDTH; y++ )
-		for( short z= 1; z< H_CHUNK_HEIGHT-1; z++ )
+		int X= k*( chunk_number_x_ - 1) * H_CHUNK_WIDTH, Y=  j * H_CHUNK_WIDTH;
+		for( int x= 0; x< H_CHUNK_WIDTH; x++ )
+		for( int y= 0; y< H_CHUNK_WIDTH; y++ )
+		for( int z= 1; z< H_CHUNK_HEIGHT-1; z++ )
 			AddSunLightSafe_r( X+x, Y+y, z, ch->SunLightLevel(x,y,z) );
 
 		const std::vector< h_LightSource* >& light_sources= ch->GetLightSourceList();
@@ -224,7 +224,7 @@ void h_World::LightWorld()
 	}
 }
 
-short h_World::RelightBlockAdd( short x, short y, short z )
+int h_World::RelightBlockAdd( int x, int y, int z )
 {
 	unsigned char l= SunLightLevel( x, y, z );
 	unsigned char fire_l= FireLightLevel( x, y, z );
@@ -233,14 +233,14 @@ short h_World::RelightBlockAdd( short x, short y, short z )
 	if( l == 0 && fire_l == 0 )
 		return 1;
 
-	short x_min, x_max, y_min, y_max, z_min, z_max;
-	x_min= x + 1 - (short)(l);
-	x_max= x - 1 + (short)(l);
-	y_min= y + 1 - (short)(l);
-	y_max= y - 1 + (short)(l);
-	z_max= ClampZ( z - 1 + (short)(l) );
+	int x_min, x_max, y_min, y_max, z_min, z_max;
+	x_min= x + 1 - (int)(l);
+	x_max= x - 1 + (int)(l);
+	y_min= y + 1 - (int)(l);
+	y_max= y - 1 + (int)(l);
+	z_max= ClampZ( z - 1 + (int)(l) );
 
-	short i, j, k;
+	int i, j, k;
 	//remove all light
 	for( i= x_min; i<= x_max; i++ )
 	for( j= y_min; j<= y_max; j++ )
@@ -252,7 +252,7 @@ short h_World::RelightBlockAdd( short x, short y, short z )
 	for( j= y_min; j<= y_max; j++ )
 	{
 		h_Chunk* ch= GetChunk( i>> H_CHUNK_WIDTH_LOG2, j>> H_CHUNK_WIDTH_LOG2 );
-		short local_i= i & ( H_CHUNK_WIDTH-1), local_j= j & ( H_CHUNK_WIDTH-1);
+		int local_i= i & ( H_CHUNK_WIDTH-1), local_j= j & ( H_CHUNK_WIDTH-1);
 		unsigned char* sun_light_map= ch->sun_light_map_ + BlockAddr( local_i, local_j, H_CHUNK_HEIGHT-2 );
 
 		unsigned int addr= BlockAddr( local_i, local_j, 0 );
@@ -290,12 +290,12 @@ short h_World::RelightBlockAdd( short x, short y, short z )
 		AddSunLight_r( i, j, z_max+1, SunLightLevel( i, j, z_max+1 ) );
 
 
-	x_min= x + 1 - (short)(fire_l);
-	x_max= x - 1 + (short)(fire_l);
-	y_min= y + 1 - (short)(fire_l);
-	y_max= y - 1 + (short)(fire_l);
-	z_min= ClampZ( z + 1 - (short)(fire_l) );
-	z_max= ClampZ( z - 1 + (short)(fire_l) );
+	x_min= x + 1 - (int)(fire_l);
+	x_max= x - 1 + (int)(fire_l);
+	y_min= y + 1 - (int)(fire_l);
+	y_max= y - 1 + (int)(fire_l);
+	z_min= ClampZ( z + 1 - (int)(fire_l) );
+	z_max= ClampZ( z - 1 + (int)(fire_l) );
 
 	//zero fire light in cube
 	for( i= x_min; i<= x_max; i++ )
@@ -328,12 +328,12 @@ short h_World::RelightBlockAdd( short x, short y, short z )
 	//shining from kight sources in cube
 	ShineFireLight( x_min, y_min, z_min, x_max, y_max, z_max );
 
-	return (short)std::max( l, fire_l );
+	return (int)std::max( l, fire_l );
 }
 
-void h_World::RelightBlockRemove( short x, short y, short z )
+void h_World::RelightBlockRemove( int x, int y, int z )
 {
-	short x1, y1, z1, z0;
+	int x1, y1, z1, z0;
 	h_Chunk* ch= GetChunk( x>> H_CHUNK_WIDTH_LOG2, y>> H_CHUNK_WIDTH_LOG2 );
 	x1= x& ( H_CHUNK_WIDTH-1);
 	y1= y& ( H_CHUNK_WIDTH-1);
@@ -379,10 +379,10 @@ void h_World::RelightBlockRemove( short x, short y, short z )
 	*/
 }
 
-void h_World::AddSunLight_r( short x, short y, short z, unsigned char l )
+void h_World::AddSunLight_r( int x, int y, int z, unsigned char l )
 {
 	h_Chunk* ch;
-	short x1, y1;
+	int x1, y1;
 	unsigned int addr;
 
 	addr= BlockAddr( x & ( H_CHUNK_WIDTH - 1 ), y & ( H_CHUNK_WIDTH - 1 ), z );
@@ -444,10 +444,10 @@ void h_World::AddSunLight_r( short x, short y, short z, unsigned char l )
 		AddSunLight_r( x1, y1, z, l1 );
 }
 
-void h_World::AddFireLight_r( short x, short y, short z, unsigned char l )
+void h_World::AddFireLight_r( int x, int y, int z, unsigned char l )
 {
 	h_Chunk* ch;
-	short x1, y1;
+	int x1, y1;
 	unsigned int addr;
 
 	addr= BlockAddr( x & ( H_CHUNK_WIDTH - 1 ), y & ( H_CHUNK_WIDTH - 1 ), z );
@@ -509,10 +509,10 @@ void h_World::AddFireLight_r( short x, short y, short z, unsigned char l )
 		AddFireLight_r( x1, y1, z, l1 );
 }
 
-void h_World::AddSunLightSafe_r( short x, short y, short z, unsigned char l )
+void h_World::AddSunLightSafe_r( int x, int y, int z, unsigned char l )
 {
 	h_Chunk* ch;
-	short x1, y1, z1;
+	int x1, y1, z1;
 	unsigned int addr;
 
 	addr= BlockAddr( x & ( H_CHUNK_WIDTH - 1 ), y & ( H_CHUNK_WIDTH - 1 ), z );
@@ -576,10 +576,10 @@ void h_World::AddSunLightSafe_r( short x, short y, short z, unsigned char l )
 		AddSunLightSafe_r( x1, y1, z, l1 );
 }
 
-void h_World::AddFireLightSafe_r( short x, short y, short z, unsigned char l )
+void h_World::AddFireLightSafe_r( int x, int y, int z, unsigned char l )
 {
 	h_Chunk* ch;
-	short x1, y1, z1;
+	int x1, y1, z1;
 	unsigned int addr;
 
 	addr= BlockAddr( x & ( H_CHUNK_WIDTH - 1 ), y & ( H_CHUNK_WIDTH - 1 ), z );
@@ -643,7 +643,7 @@ void h_World::AddFireLightSafe_r( short x, short y, short z, unsigned char l )
 		AddFireLightSafe_r( x1, y1, z, l1 );
 }
 
-void h_World::ShineFireLight( short x_min, short y_min, short z_min, short x_max, short y_max, short z_max )
+void h_World::ShineFireLight( int x_min, int y_min, int z_min, int x_max, int y_max, int z_max )
 {
 	(void)z_min;
 	(void)z_max;
@@ -653,11 +653,11 @@ void h_World::ShineFireLight( short x_min, short y_min, short z_min, short x_max
 	y_min>>= H_CHUNK_WIDTH_LOG2;
 	y_max>>= H_CHUNK_WIDTH_LOG2;
 
-	short i, j;
+	int i, j;
 	for( i= x_min; i<= x_max; i++ )
 	for( j= y_min; j<= y_max; j++ )
 	{
-		short X, Y;
+		int X, Y;
 		X= i<< H_CHUNK_WIDTH_LOG2;
 		Y= j<< H_CHUNK_WIDTH_LOG2;
 
@@ -671,12 +671,12 @@ void h_World::AddLightToBorderChunk( unsigned int X, unsigned int Y )
 {
 	h_Chunk* ch= GetChunk( X, Y );
 
-	short x= X << H_CHUNK_WIDTH_LOG2;
-	short y= Y << H_CHUNK_WIDTH_LOG2;
+	int x= X << H_CHUNK_WIDTH_LOG2;
+	int y= Y << H_CHUNK_WIDTH_LOG2;
 
-	for( short i= 0; i< H_CHUNK_WIDTH; i++ )
-	for( short j= 0; j< H_CHUNK_WIDTH; j++ )
-	for( short k= 1; k< H_CHUNK_HEIGHT-1; k++ )
+	for( int i= 0; i< H_CHUNK_WIDTH; i++ )
+	for( int j= 0; j< H_CHUNK_WIDTH; j++ )
+	for( int k= 1; k< H_CHUNK_HEIGHT-1; k++ )
 		AddSunLightSafe_r( x+i, y+j, k, ch->SunLightLevel( i, j, k ) );
 
 	//add fire lights to border chunk
@@ -710,23 +710,23 @@ void h_World::RelightWaterModifedChunksLight()
 			if( phys_processes_rand_.Rand() <=
 				phys_processes_rand_.max_rand / ( chunk_count * c_inv_desiret_chunk_update_chance ) )
 			{
-				short X= i<<H_CHUNK_WIDTH_LOG2;
-				short Y= j<<H_CHUNK_WIDTH_LOG2;
+				int X= i<<H_CHUNK_WIDTH_LOG2;
+				int Y= j<<H_CHUNK_WIDTH_LOG2;
 				ch->SunRelight();//zero light in chunk and add vertical sun light
-				for( short x= 0; x< H_CHUNK_WIDTH; x++ )
-				for( short y=0; y< H_CHUNK_WIDTH; y++ )
-				for( short z= 1; z< H_CHUNK_HEIGHT-1; z++ )
+				for( int x= 0; x< H_CHUNK_WIDTH; x++ )
+				for( int y=0; y< H_CHUNK_WIDTH; y++ )
+				for( int z= 1; z< H_CHUNK_HEIGHT-1; z++ )
 					AddSunLight_r( X+x, Y+y, z, SunLightLevel( X+x, Y+y, z) );
 
-				for( short x= 0; x< H_CHUNK_WIDTH; x++ )
-				for( short z= 1; z< H_CHUNK_HEIGHT-1; z++ )
+				for( int x= 0; x< H_CHUNK_WIDTH; x++ )
+				for( int z= 1; z< H_CHUNK_HEIGHT-1; z++ )
 				{
 					AddSunLight_r( X+x, Y-1, z, SunLightLevel( X+x, Y-1, z) );
 					AddSunLight_r( X+x, Y+H_CHUNK_WIDTH, z, SunLightLevel( X+x, Y+H_CHUNK_WIDTH, z) );
 				}
 
-				for( short y= 0; y< H_CHUNK_WIDTH; y++ )
-				for( short z= 1; z< H_CHUNK_HEIGHT-1; z++ )
+				for( int y= 0; y< H_CHUNK_WIDTH; y++ )
+				for( int z= 1; z< H_CHUNK_HEIGHT-1; z++ )
 				{
 					AddSunLight_r( X-1, Y+y, z, SunLightLevel( X-1, Y+y, z) );
 					AddSunLight_r( X+H_CHUNK_WIDTH, Y+y, z, SunLightLevel( X+H_CHUNK_WIDTH, Y+y, z) );
